refactor(bundle): name the composition root
The only description of the audit object graph lived inside
`SymfonySecurityAuditorBundle::loadExtension()` — the `config/services.php`
import, the container parameters and seven conditional registrations — so
reading or reusing the wiring meant going through a Symfony bundle.

`CoreCompositionRoot` owns it now and the bundle delegates, dropping from 394
lines to 117: the config tree and one call. It joins the `SymfonyProfile`
deptrac layer, where `CLAUDE.md` already puts the container-building classes.

This is partial progress on #250, not a close. That issue also asks the
standalone binary to build on the composition root instead of the bundle
extension, and this commit leaves that path unchanged. Doing it needs the seven
conditional registrations ported off the `ContainerConfigurator` DSL onto the
raw `ContainerBuilder` API — which forks `AttackerAgentDefinitionFactory`, since
`config/services.php` still needs the DSL form of it — and the standalone would
additionally have to run the raw config array through the config tree itself and
hand-build a `ContainerConfigurator`, across the three Symfony majors this
project supports. That is a rewrite of how the whole graph is described, and it
belongs in a change reviewed as exactly that.
