+++ deptrac.yaml
@@ -17,7 +17,7 @@ deptrac:
         - type: classLike
           value: .*\\Infrastructure\\Scan\\(IsGrantedAttributeParser|PhpParserControllerAccessControlParser|PhpParserFormBindingParser|PhpParserVoterCapabilityParser|RouteAttributeParser|SymfonyYamlSecurityConfigParser)$
         - type: classLike
-          value: .*\\Infrastructure\\Config\\(AttackerAgentDefinitionFactory|AuditConfigurationDefinition|ContainerParameterRegistrar)$
+          value: .*\\Infrastructure\\Config\\(AttackerAgentDefinitionFactory|AuditConfigurationDefinition|ContainerParameterRegistrar|CoreCompositionRoot)$
     - name: Infrastructure
       collectors:
         - type: bool
